% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions for RSA and CCA.R
\name{null.dist_RSA}
\alias{null.dist_RSA}
\title{Null Distribution for Representational Similarity Analysis (RSA)}
\usage{
null.dist_RSA(X, Y, num_permutations = 1000)
}
\arguments{
\item{X}{matrix or dataframe, a set of explanatory variables.}

\item{Y}{matrix or dataframe, a set of response variables.}

\item{num_permutations}{numeric, the number of permutations to perform. Default is 1000.}
}
\value{
Vector representing the null distribution of permuted distance correlations.
}
\description{
This function generates the null distribution of the absolute value of the distance correlations
between two sets of variables by permuting the response variables.
}
\examples{
# Example with random data
set.seed(1)
X <- matrix(rnorm(100*3), 100, 3)
Y <- matrix(rnorm(100*2), 100, 2)
null_dist_RSA(X, Y, num_permutations = 500)
}
